+++ eggs/naaya.i18n/naaya/i18n/utilities.py
@@ -18,7 +18,11 @@ class NyI18nTranslator(object):
         """
             Implementation of ITranslationDomain.translate using portal_i18n
         """
-        site = context['PARENTS'][0].getSite()
+        try:
+            site = context['PARENTS'][0].getSite()
+        except AttributeError:
+            # not in INySite context, fallback to identity translation
+            return interpolate(default or msgid, mapping)
         tool = site.getPortalI18n()
         if target_language is None:
             available = tool.get_lang_manager().getAvailableLanguages()